Reenable 2018 CDSP low power stats
Support for the CDSP master's stats was removed from 2018 device
PowerHAL code because the device is not in use and stats were not being
provided by underlying QCOM driver code.  Now that stats are being
provided, this change re-enables PowerHAL support for inclusion of CDSP
low power stats.  Even though the device is not expected to be used, the
stats should be monitor-able in order to provide a clearer picture of
device component power consumption.

Bug: 67381845
Test: Built, installed, drained power w/ flashlight, collected
bugreport, verified that CDSP stats were present in dumpsys batterystats
section.

Change-Id: Ia1dd30b6e4a4f1ecf0a9a2ef2eeae9b094819da5